Extract updato into its own component
