+++ src/components/dashboard/daybook/DayBookIndex.tsx
@@ -1,25 +1,12 @@
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  Dialog,
-  DialogTitle,
-} from '@mui/material';
+import { Card, CardContent, CardHeader } from '@mui/material';
 import { DataGrid, GridActionsCellItem, GridColDef } from '@mui/x-data-grid';
-import { FormikConfig } from 'formik';
-import { useSnackbar } from 'notistack';
 import { FC, useCallback, useMemo, useState } from 'react';
 import Iconify from 'src/components/shared/iconify';
-import {
-  useListDayBookTransactions,
-  useUpdateDayBookTransaction,
-} from 'src/hooks/cache/dayBook';
+import { useListDayBookTransactions } from 'src/hooks/cache/dayBook';
 import { getTransactionDataByDetailId } from 'src/lib/modules/dayBook';
-import { DayBookTransactionParser } from 'src/lib/parsers/dayBook';
 import { DayBookTableEntry, DayBookTransaction } from 'src/types/dayBook';
-import { toFormikValidationSchema } from 'zod-formik-adapter';
-import { DayBookTransactionForm } from './DayBookTransactionForm';
 import { DeleteDayBookTransaction } from './DeleteDayBookTransaction';
+import { UpdateDayBookTransactionForm } from './UpdateDayBookTransactionForm';
 
 const DayBookIndex: FC = () => {
   const [transactionToDelete, setTransactionToDelete] =
@@ -183,60 +170,3 @@ const DayBookIndex: FC = () => {
 };
 
 export default DayBookIndex;
-
-interface UpdateDayBookTransactionFormProps {
-  transaction: DayBookTransaction | null;
-  setTransaction: (transaction: DayBookTransaction | null) => void;
-}
-
-const UpdateDayBookTransactionForm: FC<UpdateDayBookTransactionFormProps> = ({
-  setTransaction,
-  transaction,
-}) => {
-  const handleCloseModal = () => setTransaction(null);
-  const { enqueueSnackbar } = useSnackbar();
-  const { mutateAsync: updateTransaction } = useUpdateDayBookTransaction();
-
-  const onSubmit: FormikConfig<DayBookTransaction>['onSubmit'] = async (
-    values,
-    { setSubmitting, resetForm }
-  ) => {
-    setSubmitting(true);
-
-    updateTransaction(values)
-      .then(() => {
-        resetForm();
-        handleCloseModal();
-        enqueueSnackbar('Transacción actualizada exitosamente');
-      })
-      .catch(() => {
-        enqueueSnackbar('Error al guardar la transacción', {
-          variant: 'error',
-        });
-      })
-      .finally(() => {
-        setSubmitting(false);
-      });
-  };
-
-  if (!transaction) return null;
-
-  return (
-    <Dialog
-      fullWidth
-      maxWidth="lg"
-      open={Boolean(transaction)}
-      onClose={handleCloseModal}
-    >
-      <DialogTitle>Modificar transacción</DialogTitle>
-
-      <DayBookTransactionForm
-        infoText="Aquí puedes modificar los datos de la transacción"
-        initialValues={transaction}
-        onSubmit={onSubmit}
-        onClose={handleCloseModal}
-        validationSchema={toFormikValidationSchema(DayBookTransactionParser)}
-      />
-    </Dialog>
-  );
-};
